split up lesson screen into organisms
