+++ src/client/src/components/organisms/lesson/index.ts
@@ -1,1 +1,3 @@
 export * from "./NavigationPane";
+export * from "./LessonContent";
+export * from "./InstructorPane";
